Modularize ArkheionMapView
