+++ Ascension/Modules/Arkheion/Core/ArkheionMapView+HitTesting.swift
@@ -1,0 +1,123 @@
+import SwiftUI
+
+extension ArkheionMapView {
+    // MARK: - Coordinate Mapping
+    fileprivate func convert(location: CGPoint, in geo: GeometryProxy) -> CGPoint {
+        let size = CGSize(width: geo.size.width * interactionScale,
+                          height: geo.size.height * interactionScale)
+        let origin = CGPoint(x: geo.size.width / 2 - size.width / 2,
+                             y: geo.size.height / 2 - size.height / 2)
+        return CGPoint(x: location.x + origin.x, y: location.y + origin.y)
+    }
+
+    /// Maps a raw gesture location from ``TapCaptureView`` back to the
+    /// underlying canvas space.
+    fileprivate func mapToCanvasCoordinates(location: CGPoint, in geo: GeometryProxy) -> CGPoint {
+        let center = CGPoint(x: geo.size.width / 2, y: geo.size.height / 2)
+        let currentZoom = zoom * gestureZoom
+
+        // Step 1: convert from the oversized interaction layer
+        let size = CGSize(width: geo.size.width * interactionScale,
+                          height: geo.size.height * interactionScale)
+        let origin = CGPoint(x: geo.size.width / 2 - size.width / 2,
+                             y: geo.size.height / 2 - size.height / 2)
+        var point = CGPoint(x: location.x + origin.x, y: location.y + origin.y)
+
+        // Step 2: remove pan offsets
+        point.x -= offset.width + dragTranslation.width
+        point.y -= offset.height + dragTranslation.height
+
+        // Step 3: undo zoom around the canvas center
+        point.x = center.x + (point.x - center.x) / currentZoom
+        point.y = center.y + (point.y - center.y) / currentZoom
+
+        return point
+    }
+
+    fileprivate func nearestRing(at location: CGPoint, in geo: GeometryProxy) -> Int? {
+        let center = CGPoint(x: geo.size.width / 2, y: geo.size.height / 2)
+        let point = mapToCanvasCoordinates(location: location, in: geo)
+        let distance = hypot(point.x - center.x, point.y - center.y)
+        return store.rings.min(by: { abs(distance - $0.radius) < abs(distance - $1.radius) })?.ringIndex
+    }
+
+    /// Returns the ring index and angle if the location hovers a ring edge
+    fileprivate func ringHit(at location: CGPoint, in geo: GeometryProxy) -> (Int, Double)? {
+        let center = CGPoint(x: geo.size.width / 2, y: geo.size.height / 2)
+        let point = mapToCanvasCoordinates(location: location, in: geo)
+        let distance = hypot(point.x - center.x, point.y - center.y)
+        let angle = atan2(center.y - point.y, point.x - center.x)
+        guard let ring = store.rings.min(by: { abs(distance - $0.radius) < abs(distance - $1.radius) }) else { return nil }
+        if abs(distance - ring.radius) <= 20 {
+            print("[ArkheionMap] ringHit -> index=\(ring.ringIndex) angle=\(angle)")
+            return (ring.ringIndex, Double(angle))
+        }
+
+        return nil
+    }
+
+    fileprivate func updateHover(at location: CGPoint, in geo: GeometryProxy) {
+        if let (index, angle) = ringHit(at: location, in: geo) {
+            hoverRingIndex = index
+            hoverAngle = angle
+        } else {
+            hoverRingIndex = nil
+        }
+    }
+
+    fileprivate func hitNode(at location: CGPoint, in geo: GeometryProxy) -> (branchID: UUID, nodeID: UUID)? {
+        let center = CGPoint(x: geo.size.width / 2, y: geo.size.height / 2)
+        let point = mapToCanvasCoordinates(location: location, in: geo)
+
+        for branch in store.branches {
+            guard let ring = store.rings.first(where: { $0.ringIndex == branch.ringIndex }) else { continue }
+            for (idx, node) in branch.nodes.enumerated() {
+                let distance = ring.radius + CGFloat(idx + 1) * 60
+                let position = CGPoint(
+                    x: center.x + CGFloat(Darwin.cos(branch.angle)) * distance,
+                    y: center.y + CGFloat(Darwin.sin(branch.angle)) * distance
+                )
+                let hitRadius = node.size.radius + NodeView.hitPadding
+                if hypot(point.x - position.x, point.y - position.y) <= hitRadius {
+                    print("[ArkheionMap] hitNode -> branch=\(branch.id) node=\(node.id)")
+                    return (branch.id, node.id)
+                }
+            }
+        }
+        return nil
+    }
+
+    fileprivate func hitBranch(at location: CGPoint, in geo: GeometryProxy) -> UUID? {
+        let center = CGPoint(x: geo.size.width / 2, y: geo.size.height / 2)
+        let point = mapToCanvasCoordinates(location: location, in: geo)
+
+        for branch in store.branches {
+            guard let ring = store.rings.first(where: { $0.ringIndex == branch.ringIndex }) else { continue }
+            let origin = CGPoint(
+                x: center.x + CGFloat(Darwin.cos(branch.angle)) * ring.radius,
+                y: center.y + CGFloat(Darwin.sin(branch.angle)) * ring.radius
+            )
+            let length = CGFloat(branch.nodes.count) * 60
+            let end = CGPoint(
+                x: origin.x + CGFloat(Darwin.cos(branch.angle)) * length,
+                y: origin.y + CGFloat(Darwin.sin(branch.angle)) * length
+            )
+            if distance(from: point, toSegment: origin, end: end) <= 20 {
+                print("[ArkheionMap] hitBranch -> id=\(branch.id)")
+                return branch.id
+            }
+        }
+        return nil
+    }
+
+    fileprivate func distance(from point: CGPoint, toSegment start: CGPoint, end: CGPoint) -> CGFloat {
+        let dx = end.x - start.x
+        let dy = end.y - start.y
+        let lengthSquared = dx * dx + dy * dy
+        if lengthSquared == 0 { return hypot(point.x - start.x, point.y - start.y) }
+        var t = ((point.x - start.x) * dx + (point.y - start.y) * dy) / lengthSquared
+        t = max(0, min(1, t))
+        let proj = CGPoint(x: start.x + t * dx, y: start.y + t * dy)
+        return hypot(point.x - proj.x, point.y - proj.y)
+    }
+}
